Put the ClientHttpRequestFactory in the Values object so that processors can configure it.
A consequence of this change is that ClientHttpRequestFactory  should no longer be injected into objects and instead should be obtained from the values object.
Previously several of the MapLayers performed http requests while creating the layers.  This has been changed so that the loading is done during rendering.
In order to do this Functions that take the client are passed to the MapLayer objects and they are called during rendering.

Add a processor that converts all https requests to http requests.