Add a configuration attribute for an overview map
The idea is that the configuration of an overview map has the
definition for a main map and overrides for the overview map definition.
If attributes are not set on the overview map, the defaults of the main
map will be used.

This commit also adds a stump for the overview map processor, which
internally uses the existing CreateMapProcessor.