Register Id before writing response to client.
Fix unusual case where servlet response to the client with the id and then does a context switch to another thread before registering the print ID in tempfiles map.
This allows client to make a Get print request before print id is registered.  The client will then get a 404 error that the print does not exist.