Acknowledge longitudeFirst map parameter when rendering and parsing GeoJson.
When writing a test for tiled wms I was trying to overlay GeoJson in EPSG:4326 and found that it didn't render because the features has the wrong coordinates even though longitudeFirst was missing.

First I added the parameter to the renderer so the renderer is aware of the parameter but the GeoJson features were still incorrect.  So I had to modify the GeoJSon parser to use the parameter.

The tiled-wms also has a test in core with the geojson features rendered on top.