Use Config file loader plugins for loading style files.
This commit modifies the sld style parsers so that there is just one that will try to load the style by:

* seeing if the string is sld/xml
* try to load using the ConfigFileLoaderManager
* try to load assuming the string is a URL

Also updated MapfishJsonStylePlugin so that it loads the data in the same way as the sld style plugin.

Updated FileLoaders so that they are more flexible and could be used by the style plugins for loading the style files.