Support jpg and tif external graphics.
Geotools is very strict about the graphicFormat.  It has to be one of the formats supported by ImageIO.  This is a problem at times because image/tiff or image/jpeg will be supported but image/tif or image/jpg.  Even though those are essentially aliases.
This change modifies getGraphicFormat to resolve aliases, so that geotools will accept the graphic formats.

This commit also adds a warning about DWITHIN and BEYOND