Remove requirement to add requestHeaders attribute to config.yaml.
It is added automatically if there are requestHeaders data added to the request.  This is always done by the servlet so if the request came from the servlet, then the request will have the headers in the request.

If the request comes from the commandline then it is not added and therefore is not added.

It also removes the attributes from the capabilities.