Updated the v2 api compatibility for mapfish.
In the v2 API the client can pass a url parameter which is ONLY the base url.

The current implementation assumes that the url parameter is the full request that the client made and in order to create the printUrl and createUrl and getReport urls it simply replaces the section of the url (for example dep/info.json) with the new mapping.  However this is not the behaviour needed.

This change changes keeps the current semantics but adds a check to see if the url parameter has the current handler mapping (IE dep/info.json) if not it assumes it is the base url (including the dep segment).

The concrete example was the info url used was:

http://ref.geoview.bl.ch/print3/wsgi/printproxy/info.json?url=http%3A%2F%2Fref.geoview.bl.ch%2Fprint3%2Fwsgi%2Fprintproxy

as you see the printproxy will forward to /dep/info.json.

In addition.  The old API expects the getReport url to have the form: /dep/{id}.printout.

In a previous PR I removed the .printout part not realizing that clients of the v2 API expect this for routing.  This PR adds that back.