Removed IterProcessor and replaced it with a new processor called DataSourceProcessor.
DataSourceProcessor is more general than IterPRocessors.  Previously Template had a property called IterProcessor which was a special case for making a datasource object that was put into the main table of the report object.

DataSource can do that (when the tableData property is set on the template) as well as provide datasources for sub-reports as well as for the main report