Add output value prefix to simplify fixing output mapping conflicts.
Consider the case where there are 2 CreateMapProcessors in a configuration.  CreateMapProcessors have 2 (at the moment) output values.  2 processors are not permitted to have output values with the same name.  So before this process each of the output values would have to have a mapping in the outputMapper.  This is fragile, because if a new output value is added in the future to CreateMapProcessor, then it will break the configuration because there will be 2 output values with the same name.

This PR adds an option to set a prefix for all output values with a given name.  This prefix will be applied to all output value names making them unique for the two maps.  Additionally it will be resistant to future API changes of CreateMapProcessors.