Change Configuration to lookup styles in the application context.
Move the Grid style so that it is a "named style". ie a style that is put in the application context and lookup by id.

Added a second style for the point grid (that grid type will come soon)

Added the anchor point of the label to be part of the feature (also will come in a future commit)