Add the generation of a Docker image.
The fact that we enable Docker in Travis changes the Ubuntu version used
12.04 -> 14.04(). This forced a small change in force_install.

The openjdk7 used by this new Ubuntu has a crash when the local hostname is
too long. Travis has huge hostnames and we have to change it to avoid the
crash of the JVM.

I've removed the Travis test on openjdk6 because it's not supported by
graddle-docker. Shouldn't be a problem since we still build 1.6 compatible
byte code. Nobody should use 1.6 anyway.