Correct height and width of the WMS image for high DPI
Currently, MFP doesn't take into account the DPI when calculating the
size of the image to request to the WMS server. This means that, for
instance, with a DPI of 72 and 150 it will request a tile with
WIDTH=1093 and HEIGHT=1510. The problem is that with MapServer and
MAP_RESOLUTION set to 150, the representation of the layer can be
incorrect (height and width too small for current DPI resulting in wrong
scaling of symbols). A similar issue should appear with other
cartographic servers.

With this commit, the DPI is correctly taken into account to calculate
the height and width of the tile. This means that for instance, with a
DPI of 72, it will request a tile with WIDTH=1093 and HEIGHT=1510. But
with a DPI of 150, it will request a tile with WIDTH=2277 and
HEIGHT=3146. This way, the representation of the layer is correct.

The formula used for the calculation is [based on the one used by
[MapServer](http://www.mapserver.org/development/rfc/ms-rfc-55.html#usage-example).