Fix SVG opacity.
If a layer rendered with a partial opacity contains styles with partial opacity,
we cannnot just multiply the style opacities by the the layer opacity. For example:
* background white hashed black
* layer opacity=0.7 with a line
* style1 opacity=0.5  width=10 color=green
* style2 opacity=1   width=5

We expect the line to be:
* center pure white with hints of the black hash in it
* highlight very light green with a strong presence of the black hash

If we just multiply the layer opacity to the style opacity, we end up with a
center line that is greenish.

Saddly, the generated SVG with this commit is correct (tested with chrome),
but Jasper Reports uses Batik for the rendering and Batik does not support
opacity at the `<g>` level:
https://issues.apache.org/jira/browse/BATIK-985

I've tested to set the `fill-opacity` as suggested in the issue, but it has
the same effect has the opacity muliplication trick.