Allow to add the hostname into the statsd metrics
Dropwizard's metrics are all translated into gauge in the statsd world.
Prometheus doesn't aggregate the statsg gauges, it just takes the last one.
Because of that, we just have a big mess instead of statistics when having
multiple print instances.

This patch allows to include the hostname in the statsd name by including a
`%h` in the STATSD_PREFIX environment variable. That way every instance can
publish the metrics in a different namespace.