Allow to have sub_apps and allow to have apps appear at runtime
In the application directory, you can now have directories containing
sub-directories. The app name will be {dirname_level1}:{dirname_level2}.

Also, when an unknown app is referenced in the API, the print will try to
see if the directory didn't appear in the mean time. And when a
directory disappears, remove this app from the print.

Finally, the print can now start without application configured.